changing the whole of code to use DI

+++ TetrioVirtualEnvironment/PlayerOptions.cs
@@ -2,7 +2,7 @@
 {
 	public class PlayerOptions
 	{
-		public PlayerOptions(double arr, double das, double dcd, double sdf, double safelock, bool cancel)
+		public PlayerOptions(double arr, double das, double dcd, int sdf, double safelock, bool cancel)
 		{
 			ARR = arr;
 			DAS = das;
@@ -16,7 +16,7 @@
 		public double ARR { get; internal set; }
 		public double DAS { get; internal set; }
 		public double DCD { get; internal set; }
-		public double SDF { get; internal set; }
+		public int SDF { get; internal set; }
 		public double SafeLock { get; internal set; }
 		public bool Cancel { get; internal set; }
